ParseCh introduced
Introduced an API to parse from a channel
of strings. A utility of this API is for example
drive which needs to do pre-filtering over lines
in its .driverc file at that point for which we'll
no longer have an io.Reader. It is more convenient
to just send input via a channel.